Break out io stuff into separate obj.
It's not used by the python wrapper, so no need to compile it.
