+++ tsne/bh_sne_src/Makefile
@@ -10,16 +10,19 @@ LDFLAGS += -Wl,--gc-sections
 
 all: bh_tsne
 
-bh_tsne: main.o tsne.o sptree.o
+bh_tsne: main.o io.o tsne.o sptree.o
 	$(CXX) $(CFLAGS) $(LDFLAGS) $^ -o $@
 
 sptree.o: sptree.cpp sptree.h
 	$(CXX) $(CFLAGS) -c $< -o $@
 
-main.o tsne.o: tsne.cpp tsne.h sptree.h vptree.h
+tsne.o: tsne.cpp tsne.h sptree.h vptree.h
 	$(CXX) $(CFLAGS) -c $< -o $@
 
 main.o: main.cpp tsne.h sptree.h vptree.h
+	$(CXX) $(CFLAGS) -c $< -o $@
+
+io.o: io.cpp tsne.h
 	$(CXX) $(CFLAGS) -c $< -o $@
 
 clean:
